Run script

+++ run.sh
@@ -1,3 +1,16 @@
-git fetch --all
-git reset --hard origin/master
-git pull
+cd ~/slam_path_planning
+
+# 16.04.7 LTS (Xenial Xerus)
+
+if [[grep '^VERSION="16.04.7 LTS (Xenial Xerfazswfaerus)"$' /etc/os-release]]; then
+	echo "Updating..."
+	exit
+	git fetch --all
+	git reset --hard origin/master
+	git pull
+else
+
+	echo "Not inside ROS Container; Halting"
+	exit
+
+fi
